add jpa

+++ src/main/java/org/example/demoorm/controller/CustomerController.java
@@ -63,7 +63,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}/edit")
-    public String update(@PathVariable int id, Model model) {
+    public String update(@PathVariable Long id, Model model) {
         model.addAttribute("customer", customerService.findById(id));
         return "/update";
     }
@@ -75,7 +75,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}/delete")
-    public String delete(@PathVariable int id, Model model) {
+    public String delete(@PathVariable Long id, Model model) {
         model.addAttribute("customer", customerService.findById(id));
         return "/delete";
     }
@@ -88,7 +88,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}/view")
-    public String view(@PathVariable int id, Model model) {
+    public String view(@PathVariable Long id, Model model) {
         model.addAttribute("customer", customerService.findById(id));
         return "/view";
     }
